Rename signup2 to signup
