+++ lib/screens/authentication/sign_up/components/invite_link_fail_dialog.dart
