+++ lib/screens/authentication/sign_up/mappers/create_account_state_mapper.dart
@@ -2,8 +2,8 @@ import 'package:seeds/datasource/local/models/auth_data_model.dart';
 import 'package:seeds/domain-shared/page_state.dart';
 import 'package:seeds/domain-shared/result_to_state_mapper.dart';
 import 'package:seeds/i18n/authentication/sign_up/sign_up.i18n.dart';
-import 'package:seeds/screens/authentication/sign_up2/viewmodels/page_commands.dart';
-import 'package:seeds/screens/authentication/sign_up2/viewmodels/signup_bloc.dart';
+import 'package:seeds/screens/authentication/sign_up/viewmodels/page_commands.dart';
+import 'package:seeds/screens/authentication/sign_up/viewmodels/signup_bloc.dart';
 
 class CreateAccountStateMapper extends StateMapper {
   SignupState mapResultToState(SignupState currentState, Result result, AuthDataModel authData) {
