+++ lib/screens/authentication/sign_up/viewmodels/signup_bloc.dart
@@ -8,13 +8,13 @@ import 'package:seeds/datasource/remote/model/invite_model.dart';
 import 'package:seeds/domain-shared/page_command.dart';
 import 'package:seeds/domain-shared/page_state.dart';
 import 'package:seeds/domain-shared/shared_use_cases/generate_random_key_and_words_use_case.dart';
-import 'package:seeds/screens/authentication/sign_up2/mappers/claim_invite_mapper.dart';
-import 'package:seeds/screens/authentication/sign_up2/mappers/create_account_state_mapper.dart';
-import 'package:seeds/screens/authentication/sign_up2/mappers/set_account_name_state_mapper.dart';
-import 'package:seeds/screens/authentication/sign_up2/usecases/check_account_name_availability_usecase.dart';
-import 'package:seeds/screens/authentication/sign_up2/usecases/claim_invite_usecase.dart';
-import 'package:seeds/screens/authentication/sign_up2/usecases/create_account_usecase.dart';
-import 'package:seeds/screens/authentication/sign_up2/viewmodels/page_commands.dart';
+import 'package:seeds/screens/authentication/sign_up/mappers/claim_invite_mapper.dart';
+import 'package:seeds/screens/authentication/sign_up/mappers/create_account_state_mapper.dart';
+import 'package:seeds/screens/authentication/sign_up/mappers/set_account_name_state_mapper.dart';
+import 'package:seeds/screens/authentication/sign_up/usecases/check_account_name_availability_usecase.dart';
+import 'package:seeds/screens/authentication/sign_up/usecases/claim_invite_usecase.dart';
+import 'package:seeds/screens/authentication/sign_up/usecases/create_account_usecase.dart';
+import 'package:seeds/screens/authentication/sign_up/viewmodels/page_commands.dart';
 import 'package:seeds/utils/mnemonic_code/mnemonic_code.dart';
 import 'package:seeds/utils/string_extension.dart';
